refactor: Use Go types in morse.go

+++ src/morse_test_shim.go
@@ -37,7 +37,7 @@ func morseToFile(t *testing.T, filename string, message string) {
 	audio_file_open(filename, &modem)
 	var amplitude = 100
 	gen_tone_init(&modem, amplitude, true)
-	morse_init(&modem, C.int(amplitude))
+	morse_init(&modem, amplitude)
 	morse_send(0, message, 10, 100, 100)
 	audio_file_close() // I just realised this all works on globals :s
 }
